pyproject.toml
[project]
name = "multi-agent-jarvis"
version = "0.1.0"
description = "Jarvis Multi Agent System orchestrates a modular, multi-agent architecture"
authors = [
    {name = "Sri Aradhyula", email = "sraradhy@cisco.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "prometheus-client (>=0.21.1,<0.22.0)",
    "pygithub (>=2.6.1,<3.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "werkzeug (>=3.1.3,<4.0.0)",
    "jira (>=3.8.0,<4.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "cachetools (>=5.5.2,<6.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "pyarrow (>=19.0.1,<20.0.0)",
    "matplotlib (>=3.10.1,<4.0.0)",
    "pyyaml (>=6.0.2,<7.0.0)",
    "openai (>=1.72.0,<2.0.0)",
    "gitpython (>=3.1.44,<4.0.0)",
    "pyparsing (>=3.2.3,<4.0.0)",
    "aiofiles (>=24.1.0,<25.0.0)",
    "aiohttp (>=3.11.16,<4.0.0)",
    "pydantic (>=2.11.3,<3.0.0)",
    "coloredlogs (>=15.0.1,<16.0.0)",
    "websockets (>=15.0.1,<16.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "neo4j (>=5.28.1,<6.0.0)",
    "psycopg[binary] (>=3.2.6,<4.0.0)",
    "psycopg-pool (>=3.2.6,<4.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "sse-starlette (>=2.1.3,<3.0.0)",
    "httpx-sse (>=0.4.0,<0.5.0)",
    "ruff (>=0.11.4,<0.12.0)",
    "kubernetes (==31.0.0)",
    "boto3 (>=1.37.30,<2.0.0)",
    "pyjwt-key-fetcher",
    "webexteamssdk",
    "langgraph",
    "langchain_openai",
    "langchain_community",
    "langchain_core",
    "langgraph-checkpoint-postgres",
    "asyncache",
    "langsmith",
    "backoff",
    "langgraph-cli",
    "langgraph-cli[inmem]",
]

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
]
ignore = ["F403"]


[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.poetry.scripts]
multi_agent_jarvis = "multi_agent_jarvis.agent:main"

[tool.poetry.dependencies]
uv = "^0.7.5"
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
